debugged index.html ready to push
