--- frontend/pages/Home/index.js
@@ -1,10 +1,0 @@
-const Home = () => {
-    return (
-      <div>
-        <h1>Home</h1>
-        <p>This is the Home page.</p>
-      </div>
-    );
-}
-
-export default Home;
