+++ frontend/src/App.js
@@ -1,18 +1,24 @@
-//Packages
-// import { useState } from 'react';
-// import { Routes, Route } from 'react-router-dom';
-
-//Pages
-
+import Nav from './components/Nav';
 import Home from './pages/Home';
-
-
+import LogIn from "./pages/LogIn";
+import SignUp from "./pages/SignUp";
+import NewProduct from "./pages/NewProduct";
+import EditProduct from "./pages/EditProduct";
+import { Routes, Route } from 'react-router-dom';
 
 function App() {
   return (
-    <div className="App">
-    <Home />
-    </div>
+    <main>
+      <h1>Here is my nav</h1>
+      <Nav />
+        <Routes>
+          <Route path="/home" element={<Home />} />
+          <Route path="/login" element={<LogIn />} />
+          <Route path="/signup" element={<SignUp />} />
+          <Route path="/newproduct" element={<NewProduct />} />
+          <Route path="/editproduct" element={<EditProduct />} />
+        </Routes>
+    </main>
   );
 }
