+++ frontend/src/pages/AccountPage/index.js
@@ -1,0 +1,6 @@
+const AccountPage = () => {
+    return(
+        <h1>This is our Account Page</h1>
+    )
+}
+export default AccountPage
