+++ frontend/src/pages/EditProduct/index.js
@@ -1,0 +1,6 @@
+const EditProduct = () => {
+    return(
+        <h1>This is our Edit Page</h1>
+    )
+}
+export default EditProduct
