+++ frontend/src/pages/Home/index.js
@@ -1,9 +1,10 @@
-function HomePage () {
+const Home = () => {
     return (
-        <div>
-            <h1>Home</h1>
-        </div>
-    )
+      <div>
+        <h1>Home</h1>
+        <p>This is the Home page.</p>
+      </div>
+    );
 }
 
-export default HomePage
+export default Home;
