+++ frontend/src/pages/LogIn/index.js
@@ -1,0 +1,6 @@
+const LogIn = () => {
+    return(
+        <h1>This is our Login Page</h1>
+    )
+}
+export default LogIn
