+++ frontend/src/pages/NewProduct/index.js
@@ -1,0 +1,6 @@
+const NewProduct = () => {
+    return(
+        <h1>This is our New Product Page</h1>
+    )
+}
+export default NewProduct
