+++ frontend/src/pages/ShowProduct/index.js
@@ -1,0 +1,6 @@
+const ShowProduct = () => {
+    return(
+        <h1>This is our Show Page</h1>
+    )
+}
+export default ShowProduct
